refactor(validation): NoTripleRepeat 클래스를 validation 패키지로 이동하고 import 문을 수정함
