+++ src/main/java/org/example/goormssd/usermanagementbackend/validation/NoTripleRepeatValidator.java
@@ -1,4 +1,4 @@
-package org.example.goormssd.usermanagementbackend.config.validator;
+package org.example.goormssd.usermanagementbackend.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
